Simulations: Fix some arguments and file moves
- Split every part of the command instead of keeping strings in some
  places. This requires some more chaining and formatting.
- Check if the directory already exists before trying to create it, and
  check if the files are there before moving them.

+++ simulations.py
@@ -18,18 +18,20 @@ def main(argv):
         path = '+'.join(["{}-{}".format(k,v.replace('/',',') if isinstance(v,str) else v) for (k,v) in arguments.iteritems()])
         print(path)
 
-        args = ["--{} {}".format(k.replace('_','-'), v) for (k,v) in arguments.iteritems()]
+        args = list(itertools.chain(*[["--{}".format(k.replace('_','-')), str(v)] for (k,v) in arguments.iteritems()]))
 
         with open("output.log","w") as output:
             with open("error.log","w") as error:
-                retval = subprocess.call(["python mission_basic.py"] + args + ["--no-interactive", "--location-check"], stdout=output, stderr=error)
+                retval = subprocess.call(["python", "{}/mission_basic.py".format(os.getcwd())] + args + ["--no-interactive", "--location-check"], stdout=output, stderr=error)
                 print("Return value: {}".format(retval))
 
-        os.mkdir(path)
+        if not os.path.exists(path):
+            os.mkdir(path)
 
         files = ["output.log", "error.log", "plot.eps", "map.npy"]
         for file in files:
-            shutil.move(file, path + "/" + file)
+            if os.path.exists(path):
+                shutil.move(file, path + "/" + file)
 
 if __name__ == "__main__":
     main(sys.argv[1:])
